+++ pyqir/src/context.rs
@@ -1,8 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-use crate::utils::is_all_same;
-use pyo3::{exceptions::PyValueError, prelude::*};
+use pyo3::{exceptions::PyValueError, prelude::*, PyClass};
 use std::{borrow::Borrow, ops::Deref};
 
 #[pyclass]
@@ -32,4 +31,20 @@ pub(crate) fn require_same(
     is_all_same(py, contexts)
         .then(|| ())
         .ok_or_else(|| PyValueError::new_err("Some objects come from a different context."))
+}
+
+fn is_all_same<T>(py: Python, items: impl IntoIterator<Item = impl Borrow<Py<T>>>) -> bool
+where
+    T: Eq + PyClass,
+{
+    let mut items = items.into_iter();
+    if let Some(mut prev) = items.next() {
+        for item in items {
+            if *item.borrow().borrow(py) != *prev.borrow().borrow(py) {
+                return false;
+            }
+            prev = item;
+        }
+    }
+    true
 }
